requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
matplotlib==3.7.2
numpy==1.26.1
Pillow==10.1.0
Requests==2.31.0
torch==2.1.0
torchvision==0.16.0
